+++ src/com/company/adapterPattern/AdvancedCamera.java
@@ -1,6 +1,6 @@
 package com.company.adapterPattern;
 
 public interface AdvancedCamera  {
-    public void nightScape(String fileName);
-    public void portrait(String fileName);
+    public void capture(String fileName);
+
 }
