+++ src/com/company/adapterPattern/CameraApplication.java
@@ -1,0 +1,26 @@
+package com.company.adapterPattern;
+
+class CameraApplication implements Camera {
+   CameraAdapter cameraAdapter;
+
+    @Override
+    public void capture(String fileName) {
+
+
+        if(fileName.equalsIgnoreCase("normal")){
+            System.out.println("Clicking normal picture ");
+        }
+
+
+        else if(fileName.equalsIgnoreCase("night") || fileName.equalsIgnoreCase("portrait")){
+           cameraAdapter = new CameraAdapter(fileName);
+           cameraAdapter.capture(fileName);
+
+        }
+      else{
+
+        }
+
+    }
+
+}
