+++ src/com/company/adapterPattern/NightScape.java
@@ -3,12 +3,9 @@ package com.company.adapterPattern;
  class NightScape implements AdvancedCamera{
 
     @Override
-    public void nightScape(String fileName) {
+    public void capture(String fileName) {
         System.out.println("Clicking in NightScape mode");
     }
 
-    @Override
-    public void portrait(String fileName) {
 
-    }
 }
