--- src/com/company/adapterPattern/NormalCamera.java
@@ -1,5 +1,0 @@
-package com.company.adapterPattern;
-
-public interface NormalCamera {
-    public void capture(String fileName);
-}
