+++ src/com/company/adapterPattern/Portrait.java
@@ -1,13 +1,11 @@
 package com.company.adapterPattern;
 class Portrait implements AdvancedCamera {
 
-    @Override
-    public void nightScape(String fileName) {
 
-    }
+
 
     @Override
-    public void portrait(String fileName) {
+    public void capture(String fileName) {
         System.out.println("Clicking in Portrait mode");
     }
 }
